requirements.txt
PyPDF2
pdfminer.six
argparse
tqdm